Update of axes limits, file download in release notes and integration of menu
Axes limits in the alignment plot update now whenever a new file is added to the alignment or the user changes the limits. 
An extra tab was added with information about the release and a link to download the code and associated files as .zip.
The menu is now read from the folder where the script is located.