Update QuAPPro_v0-0-5.R
App crashes when profiles are removed and then added again (especially when length or height normalization is applied again). Updating the files_to_plot within observe() instead of a reactive() seems to prevent this problem. In addition, an error message also has to be shown when a profile without total surface is added again after normalization was already performed.